refactor the unit test for the e2e gsm8k
