+++ test/registered/eval/test_text_models_gsm8k_eval.py
@@ -64,6 +64,8 @@ class TestNightlyGsm8KEval(unittest.TestCase):
             cls.models.append(ModelLaunchSettings(model_path, tp_size=2))
 
         cls.base_url = DEFAULT_URL_FOR_TEST
+        cls.env = None
+        cls.api = "sgl_eval"
 
     def test_gsm8k_all_models(self):
         warnings.filterwarnings(
@@ -85,13 +87,14 @@ class TestNightlyGsm8KEval(unittest.TestCase):
                         other_args=other_args,
                         base_url=self.base_url,
                         timeout=NIGHTLY_EVAL_SERVER_TIMEOUT,
+                        env=self.env,
                     )
 
                     args = SimpleNamespace(
                         base_url=self.base_url,
                         model=model_setup.model_path,
                         eval_name="gsm8k",
-                        api="sgl_eval",
+                        api=self.api,
                         num_examples=None,
                         num_threads=1024,
                     )
@@ -133,6 +136,14 @@ class TestNightlyGsm8KEval(unittest.TestCase):
             model_accuracy_thresholds=MODEL_SCORE_THRESHOLDS,
             model_count=len(self.models),
         )
+
+
+class TestNightlyGsm8KEvalWithRustServer(TestNightlyGsm8KEval):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.env = {"SGLANG_RUST_SERVER": "1"}
+        cls.api = "generate"
 
 
 if __name__ == "__main__":
